[notebook] Install all samples command
The goal is to support any SQL dialect and provide CREATE/INSERT
tables with also saved sample queries.

Add one iteration on the decoupling of example setups from hive
by not relying on has_concurrency() which is Hive specific only.

Still more to go but it is getting simpler/refactored with more tests.